@@ -1,5 +1,7 @@
 module github.com/cookieo9/packager
 
-go 1.16
+go 1.19
 
-require golang.org/x/tools v0.10.0
+require golang.org/x/tools v0.19.0
+
+require golang.org/x/mod v0.16.0 // indirect
